requirements.txt
aiogram
openai==1.109.1
python-dotenv